rename components called index
